requirements.txt
# ColPali Multi-Modal RAG — DSAI 413 Assignment 1
#
# For Modal deployment: pip install modal (then: modal deploy app.py)
# For local development: pip install -r requirements.txt

# Core ColPali (pinned to avoid transformers KeyError: 'qwen2_vl')
colpali-engine==0.3.9
torch>=2.5.0,<2.7.0
torchvision
transformers>=4.50.0,<4.51.0
peft>=0.14.0,<0.15.0

# PDF processing
pymupdf>=1.24.0
Pillow>=10.0.0

# Data & computation
numpy>=1.24.0
scipy>=1.11.0
scikit-learn>=1.3.0

# Visualization (notebook only)
matplotlib>=3.7.0
seaborn>=0.13.0

# Demo application
gradio>=5.0.0

# Deployment
modal>=0.64.0

# Dataset download
requests>=2.31.0
tqdm>=4.66.0

# Notebook
ipykernel>=6.25.0
ipywidgets>=8.1.0

# Optional: LLM generation (uncomment one)
# anthropic>=0.34.0        # Claude API
# google-generativeai      # Gemini API (free tier)
